moved all state to app so I could use it in the cards component to show the picked tasks
