+++ src/components/Cards.js
@@ -1,10 +1,9 @@
 import Card from "./Card.js";
-import cardsMorning from "./cards-morning.js";
 
 const Cards = (props) => {
   return (
     <div className="cards-container">
-      {cardsMorning.map((card) => {
+      {props.newMorningCards.map((card) => {
         return (
           <Card
             onClick={props.onClick}
